Fix for issue #187.
The bug was in the procedure inliner. While inlining, we were rewriting the
formal arguments with the actual arguments, but this causes weird problems if a
local variable of the procedure has the same name as one of the actual
arguments. The fix we have now creates a bunch of new temporary variables to
hold the actual arguments. These new variables are named __ucld_arg_<something>
and the actual arguments are copied into these temp variables first. We then
rewrite the formal arguments to these temp variables.