Fix + tests for #188.
We were only checking for recursion in procedures actually being called from
init/next. In combination with recent changes to the ProcedureInliner, this
could result in a stack overflow when there was recursion in such "unreachable"
procedures.

Note these procedures may not be really be unreachable thanks to instance
procedure calls.